Move vector.h to old_vector.h and add a shim vector.h To do so, one instance of the class keyword needs to be removed.
